requirements.txt
google-adk>=0.3.0
paper-search-mcp
